# swagger.yaml
swagger: '2.0'
info:
  version: v4
  title: Projects API
host: 'localhost:3000'
basePath: /v4
schemes:
  - http
produces:
  - application/json
consumes:
  - application/json
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
paths:
...
  '/projects/{projectId}/phases':
    parameters:
      - $ref: '#/parameters/projectIdParam'
    get:
      tags:
        - phase
      operationId: findProjectPhases
      security:
        - Bearer: []
      description: >-
        Retrieve all project phases. All users who can edit project can access
        this endpoint.
      parameters:
        - name: fields
          required: false
          type: string
          in: query
          description: |
            Comma separated list of project phase fields to return.
        - name: sort
          required: false
          description: >
            sort project phases by startDate, endDate, status, order. Default is
            startDate asc
          in: query
          type: string
      responses:
        '200':
          description: A list of project phases
          schema:
            $ref: '#/definitions/ProjectPhaseListResponse'
        '403':
          description: No permission or wrong token
          schema:
            $ref: '#/definitions/ErrorModel'
    post:
      tags:
        - phase
      operationId: addProjectPhase
      security:
        - Bearer: []
      description: >-
        Create a project phase. It also updates the `order` field of all other
        phases in the same project which have `order` greater than or equal to
        the `order` specified in the POST body.
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            allOf:
              - $ref: '#/definitions/ProjectPhaseBodyParam'
            properties:
              param:
                type: object
                properties:
                  productTemplateId:
                    type: number
                    format: long
                    description: >-
                      the optional productTemplateId used to populate a new
                      phase product for the created phase
      responses:
        '201':
          description: Returns the newly created project phase
          schema:
            $ref: '#/definitions/ProjectPhaseResponse'
        '403':
          description: No permission or wrong token
          schema:
            $ref: '#/definitions/ErrorModel'
        '422':
          description: Invalid input
          schema:
            $ref: '#/definitions/ErrorModel'
  '/projects/{projectId}/phases/{phaseId}':
    parameters:
      - $ref: '#/parameters/projectIdParam'
      - $ref: '#/parameters/phaseIdParam'
    get:
      tags:
        - phase
      description: >-
        Retrieve project phase by id. All users who can edit project can access
        this endpoint.
      security:
        - Bearer: []
      responses:
        '200':
          description: a project phase
          schema:
            $ref: '#/definitions/ProjectPhaseResponse'
        '403':
          description: No permission or wrong token
          schema:
            $ref: '#/definitions/ErrorModel'
        '404':
          description: Not found
          schema:
            $ref: '#/definitions/ErrorModel'
      parameters:
        - $ref: '#/parameters/phaseIdParam'
      operationId: getProjectPhase
    patch:
      tags:
        - phase
      operationId: updateProjectPhase
      security:
        - Bearer: []
      description: >-
        Update a project phase. All users who can edit project can access this
        endpoint. It also updates the `order` field of all other phases in the
        same project which have `order` greater than or equal to the `order`
        specified in the POST body.
      responses:
        '200':
          description: Successfully updated project phase.
          schema:
            $ref: '#/definitions/ProjectPhaseResponse'
        '403':
          description: No permission or wrong token
          schema:
            $ref: '#/definitions/ErrorModel'
        '404':
          description: Not found
          schema:
            $ref: '#/definitions/ErrorModel'
        '422':
          description: Invalid input
          schema:
            $ref: '#/definitions/ErrorModel'
        default:
          description: error payload
          schema:
            $ref: '#/definitions/ErrorModel'
      parameters:
        - $ref: '#/parameters/phaseIdParam'
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/ProjectPhaseBodyParam'
    delete:
      tags:
        - phase
      description: >-
        Remove an existing project phase. All users who can edit project can
        access this endpoint.
      security:
        - Bearer: []
      parameters:
        - $ref: '#/parameters/phaseIdParam'
      responses:
        '204':
          description: Project phase successfully removed
        '403':
          description: No permission or wrong token
          schema:
            $ref: '#/definitions/ErrorModel'
        '404':
          description: If project is not found
          schema:
            $ref: '#/definitions/ErrorModel'
...
parameters:
  projectIdParam:
    name: projectId
    in: path
    description: project identifier
    required: true
    type: integer
    format: int64
  phaseIdParam:
    name: phaseId
    in: path
    description: project phase identifier
    required: true
    type: integer
    format: int64
    minimum: 1
...
definitions:
  ResponseMetadata:
    title: Metadata object for a response
    type: object
    properties:
      totalCount:
        type: integer
        format: int64
        description: Total count of the objects
  ErrorModel:
    type: object
    properties:
      id:
        type: string
        description: unique id identifying the request
      version:
        type: string
      result:
        type: object
        properties:
          success:
            type: boolean
          status:
            description: http status code
            type: integer
            format: int32
          debug:
            type: object
          content:
            type: object
...
  ProjectPhaseRequest:
    title: Project phase request object
    type: object
    required:
      - name
      - status
      - startDate
      - endDate
    properties:
      name:
        type: string
        description: the project phase name
      description:
        type: string
        description: the project phase short description
      requirements:
        type: string
        description: the project phase requirements
      status:
        type: string
        description: the project phase status
      startDate:
        type: string
        format: date
        description: the project phase start date
      endDate:
        type: string
        format: date
        description: the project phase end date
      budget:
        type: number
        description: the project phase budget
      progress:
        type: number
        description: the project phase progress
      details:
        type: object
        description: the project phase details
      order:
        type: number
        format: integer
        description: the project phase order
...
  ProjectPhaseBodyParam:
    title: Project phase body param
    type: object
    required:
      - param
    properties:
      param:
        $ref: '#/definitions/ProjectPhaseRequest'
  ProjectPhase:
    title: Project phase object
    allOf:
      - type: object
        required:
          - id
          - createdAt
          - createdBy
          - updatedAt
          - updatedBy
        properties:
          id:
            type: number
            format: int64
            description: the id
          createdAt:
            type: string
            description: Datetime (GMT) when object was created
            readOnly: true
          createdBy:
            type: integer
            format: int64
            description: READ-ONLY. User who created this object
            readOnly: true
          updatedAt:
            type: string
            description: READ-ONLY. Datetime (GMT) when object was updated
            readOnly: true
          updatedBy:
            type: integer
            format: int64
            description: READ-ONLY. User that last updated this object
            readOnly: true
      - $ref: '#/definitions/ProjectPhaseRequest'
...
  ProjectPhaseResponse:
    title: Single project phase response object
    type: object
    properties:
      id:
        type: string
        description: unique id identifying the request
      version:
        type: string
      result:
        type: object
        properties:
          success:
            type: boolean
          status:
            type: integer
            format: int32
            description: http status code
          metadata:
            $ref: '#/definitions/ResponseMetadata'
          content:
            $ref: '#/definitions/ProjectPhase'
  ProjectPhaseListResponse:
    title: Project phase list response object
    type: object
    properties:
      id:
        type: string
        readOnly: true
        description: unique id identifying the request
      version:
        type: string
      result:
        type: object
        properties:
          success:
            type: boolean
          status:
            type: integer
            format: int32
            description: http status code
          metadata:
            $ref: '#/definitions/ResponseMetadata'
          content:
            type: array
            items:
              $ref: '#/definitions/ProjectPhase'
...